pyproject.toml
[tool.poetry]
name = "frizbot"
version = "0.1.0"
description = "RobotFramework experiments"
authors = ["dhilipsiva <dhilipsiva@pm.me>"]

[tool.poetry.dependencies]
Faker = "^9.7.1"
docutils = "^0.17.1"
faker-vehicle = "^0.2.0"
python = "^3.9"
robotframework = "^4.1.1"
robotframework-jsonlibrary = "^0.3.1"
robotframework-jsonvalidator = "^2.0.0"
robotframework-requests = "^0.9.2"
sa11y = "^0.2.3"
jq = "^1.2.1"
bzt = "^1.16.2"
robotframework-browser = "^13.0.0"
robotframework-pdf2textlibrary = "^1.0.1"
PyPDF2 = "^2.11.0"
robotframework-SikuliLibrary = "^2.0.3"
robotframework-imagehorizonlibrary = "^1.0"
opencv-python = "^4.6.0"
robotframework-xvfb = "^1.2.2"
xvfbwrapper = "0.2.5"
PyVirtualDisplay = "^3.0"
robotframework-imaplibrary = "^0.3.0"
robotframework-pabot = "^2.7.0"

[tool.poetry.dev-dependencies]
neovim = "^0.3.1"
ipython = "^7.27.0"
flake8 = "^3.9.2"
pep8 = "^1.7.1"
ipdb = "^0.13.9"
black = "^21.9b0"
isort = "^5.9.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
# Make sure this matches `*.py` in .editorconfig
ensure_newline_before_comments = true
force_grid_wrap = 0
include_trailing_comma = true
line_length = 88
multi_line_output = 3
use_parentheses = true
